requirements.txt
Flask
numpy
PyAudio
librosa
matplotlib
torch
python-rtmidi
numba